feat: complete CoachOverlay 7-state presence layer + voice pipeline improvements + spirit-coach architecture upgrades
- CoachOverlay: add same-state refresh for bubble states (hint/speaking/thinking)
- CoachClient: add exponential backoff reconnect (1s→30s, max 10 attempts)
- VoicePipeline: add min speech duration (500ms), voice cooldown (3s), mute mic feedback
- HybridAPI: add error UI notification, HTTP response code validation, int timestamp fix
- spirit-coach-service: replace setInterval with async consumeLoop, add stale session cleanup
- voice-service: add F5-TTS voice mapping, update ElevenLabs default voice
- Godot client: add SpiritForest scene, MagicFlower/TreasureChest components
- Tests: comprehensive updates across 4 test files (+720/-130 lines)
- Fix: docker-compose REDIS_URL port 6379, coach-ws connection interface

+++ services/spirit-coach-service/src/types/coach-events.ts
@@ -14,7 +14,7 @@ const dialogueTurnSchema = z.object({
   player_text: z.string().min(1),
   npc_response: z.string().min(1),
   language: z.string().min(1),
-  timestamp: z.number().int(),
+  timestamp: z.number(),
 })
 
 const silenceTimeoutSchema = z.object({
@@ -23,7 +23,7 @@ const silenceTimeoutSchema = z.object({
   user_id: z.string().min(1),
   npc_id: z.string().min(1),
   silence_ms: z.number().int().min(15000),
-  timestamp: z.number().int(),
+  timestamp: z.number(),
 })
 
 const wakeRequestSchema = z.object({
@@ -32,7 +32,7 @@ const wakeRequestSchema = z.object({
   user_id: z.string().min(1),
   npc_id: z.string().min(1),
   player_text: z.string().min(1),
-  timestamp: z.number().int(),
+  timestamp: z.number(),
 })
 
 export const coachInputSchema = z.union([
@@ -52,7 +52,7 @@ export const coachInterventionSchema = z.object({
   emotion: z.string().min(1),
   should_tts: z.boolean(),
   ttl_ms: z.number().int().positive(),
-  timestamp: z.number().int(),
+  timestamp: z.number(),
 })
 
 export type CoachInput = z.infer<typeof coachInputSchema>
